refactor scrollprogress prevent about rerender
